Moved seprate components from gosub_engine to gosub_net
